--- Content.Client/Chat/TypingIndicator/TypingIndicatorClothingSystem.cs
@@ -1,7 +1,0 @@
-﻿using Content.Shared.Chat.TypingIndicator;
-
-namespace Content.Client.Chat.TypingIndicator;
-
-public sealed class TypingIndicatorClothingSystem : SharedTypingIndicatorClothingSystem
-{
-}
