+++ Content.Shared/Chat/TypingIndicator/TypingIndicatorComponent.cs
@@ -8,7 +8,7 @@ namespace Content.Shared.Chat.TypingIndicator;
 ///     Added automatically when player poses entity.
 /// </summary>
 [RegisterComponent, NetworkedComponent]
-[Access(typeof(SharedTypingIndicatorSystem), typeof(SharedTypingIndicatorClothingSystem))]
+[Access(typeof(SharedTypingIndicatorSystem))]
 public sealed class TypingIndicatorComponent : Component
 {
     /// <summary>
